DevIchAc97: warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@54422 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DevIchAc97.cpp
@@ -1195,7 +1195,7 @@ static int ichac97ReadAudio(PAC97STATE pThis, PAC97BMREG pReg, uint32_t cbMax, u
     uint32_t cbRead = 0;
 
     size_t cbMixBuf = cbMax;
-    uint32_t cbToRead = RT_MIN(pReg->picb << 1, cbMixBuf);
+    uint32_t cbToRead = RT_MIN((uint32_t)(pReg->picb << 1), cbMixBuf);
 
     if (!cbToRead)
     {
